requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
alembic=1.13.1=pypi_0
aniso8601=9.0.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asgiref=3.8.1=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.1.2=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
blinker=1.8.2=pypi_0
build=1.2.1=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.5.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.14=pypi_0
docker=7.1.0=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastapi=0.110.3=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.14.0=pypi_0
flasgger=0.9.7.1=pypi_0
flask=3.0.3=pypi_0
flask-cors=4.0.1=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.51.0=pypi_0
fqdn=1.5.1=pypi_0
fsspec=2024.3.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.29.0=pypi_0
googleapis-common-protos=1.63.0=pypi_0
graphene=3.3=pypi_0
graphql-core=3.2.3=pypi_0
graphql-relay=3.2.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.63.0=pypi_0
gunicorn=22.0.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipykernel=6.29.4=pypi_0
ipython=8.24.0=pypi_0
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
json5=0.9.25=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.1.8=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.1=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
kubernetes=29.0.0=pypi_0
lab=8.1=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
mako=1.3.5=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mlflow=2.14.1=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nltk=3.8.1=pypi_0
notebook=7.1.3=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
ollama=0.1.9=pypi_0
onnxruntime=1.17.3=pypi_0
openssl=3.0.13=h7f8727e_1
opentelemetry-api=1.24.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.24.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.24.0=pypi_0
opentelemetry-instrumentation=0.45b0=pypi_0
opentelemetry-instrumentation-asgi=0.45b0=pypi_0
opentelemetry-instrumentation-fastapi=0.45b0=pypi_0
opentelemetry-proto=1.24.0=pypi_0
opentelemetry-sdk=1.24.0=pypi_0
opentelemetry-semantic-conventions=0.45b0=pypi_0
opentelemetry-util-http=0.45b0=pypi_0
orjson=3.10.2=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py310h06a4308_0
platformdirs=4.2.1=pypi_0
posthog=3.5.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
psycopg2-binary=2.9.9=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=15.0.2=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pypdf2=3.0.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
querystring-parser=1.2.4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rouge-score=0.1.2=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
send2trash=1.8.3=pypi_0
setuptools=68.2.2=py310h06a4308_0
shellingham=1.5.4=pypi_0
simplejson=3.19.2=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.30=pypi_0
sqlite=3.45.3=h5eee18b_0
sqlparse=0.5.0=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.37.2=pypi_0
style=1.1.0=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.41.2=pypi_0
txt2tags=3.9=pypi_0
typer=0.12.3=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
update=0.0.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.21.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.10=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_1
